Eliminate logspam using filelog monitor
When using a filelog monitor, the top-level `pluginConfig.message` filter provides a limited stream of log events to be checked against the `rules[].pattern`.  I thus expect it to be normal for most log events to fail to match the top-level `pluginConfig.message` filter. Such a condition should not trigger a warning-level log message.  And in fact such log messages should be suppressed by default, unless `-v=5` or higher is used for troubleshooting/debug purposes.